pyproject.toml
[tool.poetry]
name = "djangosampleproject"
version = "0.1.0"
description = ""
authors = ["Beki Imanji <imanjibeki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.14"
djangorestframework = "3.15.2"
django-split-settings = "1.3.2"
python-decouple = "3.8"
psycopg2-binary = "2.9.9"
gunicorn = "22.0.0"
drf-yasg = "1.21.7"
bcrypt = "4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
